[localevaluation.elliptic] add safe integer conversion

+++ dune/gdt/localevaluation/elliptic.hh
@@ -10,6 +10,8 @@
 
 #include <tuple>
 
+#include <boost/numeric/conversion/cast.hpp>
+
 #include <dune/common/dynmatrix.hh>
 #include <dune/common/typetraits.hh>
 
@@ -170,8 +172,8 @@ private:
       const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rT, rCT>& testBase,
       const Stuff::LocalfunctionSetInterface<EntityType, DomainFieldType, dimDomain, R, rA, rCA>& ansatzBase) const
   {
-    return localFunction.order() + std::max(ssize_t(testBase.order()) - 1, ssize_t(0))
-           + std::max(ssize_t(ansatzBase.order()) - 1, ssize_t(0));
+    return localFunction.order() + boost::numeric_cast<size_t>(std::max(ssize_t(testBase.order()) - 1, ssize_t(0)))
+           + boost::numeric_cast<size_t>(std::max(ssize_t(ansatzBase.order()) - 1, ssize_t(0)));
   } // size_t redirect_order( ... )
 
   /**
